Add page metadata capture and tests
