+++ .gitignore
@@ -2,6 +2,8 @@
 mlruns/
 **/mlruns/
 
+tests/output
+
 
 # Bundle
 .databricks/
