--- 02b_convert_databricks.py
@@ -1,9 +1,0 @@
-# Databricks notebook source
-# MAGIC %md
-# MAGIC # Convert (Databricks)
-# MAGIC
-# MAGIC Databricks leverages the power of vision language models (VLMs) to provide a scalable native solution for document extraction.
-
-# COMMAND ----------
-
-print("Hello, Databricks")
