+++ maud/document/converters.py
@@ -1,18 +1,53 @@
-from abc import ABC, abstractmethod
-import os
 import hashlib
 import json
 from pathlib import Path
-from datetime import datetime, timedelta
 from docling.document_converter import DocumentConverter
 from docling.datamodel.document import DoclingDocument
-from docling_core.types.doc import ImageRefMode
+from docling_core.types.doc import ImageRefMode, PageItem
 import logging
+from docling.datamodel.pipeline_options import PdfPipelineOptions
+from maud.document.metadata import MetaDataType, DescriptionData
+from typing import Dict
 
 
-class AbstractConverter(ABC):
-    def __init__(self, input_path: Path, output_dir: Path, overwrite=False):
-        self.doc_file_name = "doc.md"
+class ExtendedDocument(DoclingDocument):
+    page_metadata: Dict[int, MetaDataType] = {}
+
+    class Config:
+        extra = "allow"
+
+
+class PageMetadataModel:
+    """
+    A model that can be used to describe a page and extract hierarchy
+    """
+
+    def __init__(self, enabled: bool = True):
+        self.enabled = enabled
+
+    def analyze_pages(self, page_batch: Dict[int, PageItem]) -> Dict[int, MetaDataType]:
+        if not self.enabled:
+            return {}
+
+        page_metadata = {}
+        for idx, page in page_batch.items():
+            assert isinstance(page, PageItem)
+            page_metadata[idx] = DescriptionData(
+                provenance="a llm model",
+                text="This is a dummy description",
+            )
+
+        if not page_metadata:
+            return {0: DescriptionData(provenance="", text="No pages in document")}
+
+        return page_metadata
+
+
+class MaudConverter(DocumentConverter):
+    def __init__(self, input_path: Path, output_dir: Path, overwrite=False, **kwargs):
+        super().__init__(**kwargs)
+        self.doc_file_name = "doc.json"
+        self.md_file_name = "doc.md"
         self.input_path = input_path
         self.output_dir = output_dir
         self.overwrite = overwrite
@@ -52,64 +87,21 @@ class AbstractConverter(ABC):
 
         return True
 
-    @abstractmethod
-    def convert(self, overwrite=False):
-        self.logger.info("Converting document")
-        """
-        Convert the file into a structured document.
-        """
-        pass
-
-    @abstractmethod
-    def add_descriptions(self):
-        self.logger.info("Adding descriptions to the document")
-        """
-        Add descriptions to the document using a multimodal model..
-        """
-        pass
-
-    @abstractmethod
-    def save_document(self):
-        self.logger.info("Saving document")
-        """
-        Save the document to the output directory in a reloadable format .
-        """
-        pass
-
-    @abstractmethod
-    def load_document(self):
-        self.logger.info("Loading document")
-        """
-        Reload the document from the output directory.
-        """
-        pass
-
-
-class DoclingConverterAdapter(AbstractConverter):
-    def __init__(self, input_path: Path, output_dir: Path, overwrite=False, **kwargs):
-        super().__init__(input_path, output_dir, overwrite)
-        self.converter = DocumentConverter(**kwargs)
-        self.doc_file_name = "doc.json"
-        self.md_file_name = "doc.md"
-
-    def convert(self):
+    def convert(self, *args, **kwargs):
         self.logger.info("Converting document")
 
         if self._validate_output_exists():
-            self.logger.info("Conversion exists, reloading")
-            self.result = self.load_result()
-            return self.result
+            self.logger.info("Conversion exists, reloading document")
+            self.document = self.load_document()
 
-        self.result = self.converter.convert(self.input_path)
+        self.result = super().convert(self.input_path, *args, **kwargs)
+
+        self.result.document = ExtendedDocument(
+            **self.result.document.model_dump(),
+            page_metadata=PageMetadataModel().analyze_pages(self.result.document.pages),
+        )
+
         self.document = self.result.document
-        return self.document
-
-    def add_descriptions(self):
-        self.logger.info("Adding descriptions to the document")
-        """
-        Add descriptions to the document using a multimodal model.
-        """
-        pass
 
     def load_document(self):
         self.logger.info("Loading document")
@@ -117,13 +109,12 @@ class DoclingConverterAdapter(AbstractConverter):
         with (self._output_path / self.doc_file_name).open("r") as fp:
             doc_dict = json.loads(fp.read())
 
-        self.document = DoclingDocument.model_validate(doc_dict)
-        return self.document
+        self.document = ExtendedDocument.model_validate(doc_dict)
 
     def save_document(self):
         self.logger.info("Saving document")
 
-        self._output_path.mkdir(parents=False, exist_ok=True)
+        self._output_path.mkdir(parents=True, exist_ok=True)
 
         self.document.save_as_markdown(
             self._output_path / self.md_file_name, image_mode=ImageRefMode.EMBEDDED
@@ -133,4 +124,21 @@ class DoclingConverterAdapter(AbstractConverter):
             self._output_path / self.doc_file_name, image_mode=ImageRefMode.EMBEDDED
         )
 
-        return True
+
+class MAUDPipelineOptions(PdfPipelineOptions):
+
+    images_scale: float = 2.0
+    # pages
+    generate_page_images: bool = True
+    describe_pages: bool = True
+    classify_pages: bool = True
+
+    # pictures
+    generate_picture_images: bool = True
+    describe_pictures: bool = True
+    classify_pictures: bool = True
+
+    # tables
+    generate_table_images: bool = True
+    describe_tables: bool = True
+    classify_tables: bool = True
