+++ pyproject.toml
@@ -63,5 +63,10 @@ check_untyped_defs = true
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
+pythonpath = "maud"
 python_files = ["test_*.py"]
 addopts = "--cov=maud --cov-report=term-missing" 
+markers = [
+    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
+    "serial",
+]
