--- tests/test_agent_functions.py
@@ -1,5 +1,0 @@
-from maud.agent.functions import add
-
-
-def test_add():
-    assert add(1, 2) == 3
